pyproject.toml
[tool.poetry]
name = "testquest"
version = "0.1.0"
description = ""
authors = ["Dmitry Veko <blankvektrius@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^22.6.2"
sanic-ext = "^22.6.3"
sanic-jwt = "^1.8.0"
SQLAlchemy = "^1.4.40"
asyncpg = "^0.26.0"
pycryptodome = "^3.15.0"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
